Made kitchen window a fractured geometry
You need to make the damage values for when the GeometryCollection smashes to something very low like 1, and then make sure Simulate Physics is turned off, and only once you want the glass to break, turn Simulate Physics on to instantly break it.

The implementation is weird because Unreal doesn't really work correctly with fractures and things.