Added door SMs
Wardrobe and kitchen cupboard doors are now static meshes. I have not had the time to go and actually make them animate, but you should be able to stick an animator or sequence onto each one to move the doors to a specific angle you want. Because they are now static meshes, its easier to do.